requirements.txt
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
squareup==26.0.0.20230816
flask==3.0.0
requests==2.31.0
python-dotenv==1.0.0
